+++ packages/stream_chat/lib/stream_chat.dart
@@ -82,6 +82,7 @@ export 'src/client/channel/channel_read_helper.dart';
 export 'src/client/channel_delivery_reporter.dart';
 export 'src/client/client.dart';
 export 'src/client/key_stroke_handler.dart';
+export 'src/client/moderation_client.dart';
 export 'src/client/query_channels_result.dart';
 export 'src/client/retry_policy.dart';
 export 'src/core/api/attachment_file_uploader.dart';